turn flaw validate into a generic ValidationMixin
